my_chat_analyzer/requirements.txt
fastapi
uvicorn
python-multipart
jinja2
google-generativeai
python-dotenv